pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "pdfplumber>=0.11.7",
    "streamlit>=1.47.0",
]
